Fix a problem with vsphere machinesets on ocp 4.19
OCP 4.19 has machine sets with a different template name.  We need to
switch to the template name being used since the current one does not
exist on this level.
I also included some files that will be useful for dryrun testing of
this policy which contains a lot of templating.

Refs:
 - https://issues.redhat.com/browse/OCPQE-29046

Signed-off-by: Gus Parvin <gparvin@redhat.com>